Support for bastion/boot node concept

+++ scripts/09-0-kubeconfig.sh
@@ -4,7 +4,7 @@ source 00-variables.sh
 
 set -e
 
-cd /opt/ibm-cloud-private-${INCEPTION_VERSION}
+cd $(dirname "$(find /opt -path "*" -type d -name "cluster")")
 
 # Get kubectl
 # Append a -ee on the version number for Cloud Native Installs (e.g. v1.9.1-ee)
@@ -12,9 +12,29 @@ sudo docker run -e LICENSE=accept --net=host -v /usr/local/bin:/data ibmcom/icp-
 
 # Make config directory
 mkdir -p ~/.kube
-sudo cp /var/lib/kubelet/kubectl-config ~/.kube/config
-sudo cp /etc/cfc/conf/kubecfg.crt ~/.kube/kubecfg.crt
-sudo cp /etc/cfc/conf/kubecfg.key ~/.kube/kubecfg.key
+
+
+if [ ! -f /var/lib/kubelet/kubectl-config ]; then
+  # Boot (Bastion) Node approach
+  scp ${SSH_USER}@$MASTER_IP:/var/lib/kubelet/kubectl-config ~/.kube/config
+else
+  sudo cp /var/lib/kubelet/kubectl-config ~/.kube/config
+fi
+
+if [ ! -f /etc/cfc/conf/kubecfg.crt ]; then
+  # Boot (Bastion) Node approach
+  scp ${SSH_USER}@$MASTER_IP:/etc/cfc/conf/kubecfg.crt ~/.kube/kubecfg.crt
+else
+  sudo cp /etc/cfc/conf/kubecfg.crt ~/.kube/kubecfg.crt
+fi
+
+if [ ! -f /etc/cfc/conf/kubecfg.key ]; then
+  # Boot (Bastion) Node approach
+  scp ${SSH_USER}@$MASTER_IP:/etc/cfc/conf/kubecfg.key ~/.kube/kubecfg.key
+else
+  sudo cp /etc/cfc/conf/kubecfg.key ~/.kube/kubecfg.key
+fi
+
 sudo chown -R $USER  ~/.kube/
 
 #Set kube config
